+++ packages/dev/core/src/XR/webXRLayerWrapper.ts
@@ -5,9 +5,10 @@ import type { WebXRSessionManager } from "./webXRSessionManager";
 /** Covers all supported subclasses of WebXR's XRCompositionLayer */
 // TODO (rgerd): Extend for all other subclasses of XRCompositionLayer.
 export type WebXRCompositionLayerType = "XRProjectionLayer";
+export type WebXRQuadLayerType = "XRQuadLayer";
 
 /** Covers all supported subclasses of WebXR's XRLayer */
-export type WebXRLayerType = "XRWebGLLayer" | WebXRCompositionLayerType;
+export type WebXRLayerType = "XRWebGLLayer" | WebXRCompositionLayerType | WebXRQuadLayerType;
 
 /**
  * Wrapper over subclasses of XRLayer.
